Module5 mvc

+++ src/main/java/com/company/service/impl/BookServiceImpl.java
@@ -1,23 +1,39 @@
 package com.company.service.impl;
 
 import com.company.dao.entity.Book;
-import com.company.repository.impl.BookRepositoryImpl;
+import com.company.repository.BookRepository;
 import com.company.service.BookService;
+import com.company.service.exception.ApplicationNotFoundException;
+import com.company.service.exception.ApplicationValidationException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
-import java.math.BigDecimal;
 import java.util.List;
 import java.util.Objects;
 
 @Service
 public class BookServiceImpl implements BookService {
-
-    private final BookRepositoryImpl bookRepository;
+    private final BookRepository bookRepository;
 
     @Autowired
-    public BookServiceImpl(BookRepositoryImpl bookRepository) {
+    public BookServiceImpl(BookRepository bookRepository) {
         this.bookRepository = bookRepository;
+    }
+
+    @Override
+    public Book findById(Long id) {
+        Book book = bookRepository.findById(id);
+        if (book == null) {
+            throw new ApplicationNotFoundException("Can't find book with id=" + id);
+        }
+        return book;
+    }
+
+    @Override
+    public Book create(Book book) {
+        Book newBook = bookRepository.create(book);
+        validateIsbn(book);
+        return newBook;
     }
 
     @Override
@@ -26,44 +42,11 @@ public class BookServiceImpl implements BookService {
     }
 
     @Override
-    public Book findById(Long id) {
-        Book book = bookRepository.findById(id);
-        if (book == null) {
-            throw new RuntimeException("Can't find book with id=" + id);
-        }
-        return book;
-    }
-
-    @Override
     public void delete(Long id) {
-        boolean success = bookRepository.delete(id);
-        if (!success) {
-            throw new RuntimeException("Can't find book with id=" + id);
-        }
-    }
-
-    @Override
-    public Book create(Book entity) {
-        Book book = bookRepository.create(entity);
-        validatePrice(book);
-        validateIsbn(entity);
-        return book;
-    }
-
-    @Override
-    public Book update(Book entity) {
-        Book book = bookRepository.update(entity);
-        validatePrice(book);
-        validateIsbn(entity);
-        if (book == null) {
-            throw new RuntimeException("Can't find book with");
-        }
-        return book;
-    }
-
-    public void validatePrice(Book book) {
-        if (book.getPrice().compareTo(BigDecimal.ZERO) < 0) {
-            throw new RuntimeException("Price is not valid");
+        if (!bookRepository.delete(id)) {
+            throw new ApplicationNotFoundException("No book with id: " + id);
+        } else {
+            bookRepository.delete(id);
         }
     }
 
@@ -71,7 +54,7 @@ public class BookServiceImpl implements BookService {
         List<Book> books = findAll();
         for (int j = 0; j < books.size() - 1; j++) {
             if (books.get(j) != null && Objects.equals(books.get(j).getIsbn(), entity.getIsbn())) {
-                throw new RuntimeException("Book with isbn: " + entity.getIsbn() + " already exist");
+                throw new ApplicationValidationException("Book with isbn: " + entity.getIsbn() + " already exist");
             }
         }
     }
